More explicit conversions

+++ include/SoloRenderTarget.h
@@ -19,10 +19,10 @@ namespace solo
         virtual void bind() = 0;
         virtual void unbind() = 0;
 
-        virtual void setColorAttachment(size_t index, shared<Texture2D> texture) = 0;
+        virtual void setColorAttachment(int index, shared<Texture2D> texture) = 0;
         virtual Vector2 getColorAttachmentSize() const = 0;
-        virtual size_t getColorAttachmentCount() const = 0;
-        virtual shared<Texture2D> getColorAttachment(size_t index) const = 0;
+        virtual int getColorAttachmentCount() const = 0;
+        virtual shared<Texture2D> getColorAttachment(int index) const = 0;
 
     protected:
         RenderTarget() {}
