+++ src/platform/opengl/SoloOpenGLRenderTarget.h
@@ -15,10 +15,10 @@ namespace solo
         virtual void bind() override;
         virtual void unbind() override;
 
-        virtual void setColorAttachment(size_t index, shared<Texture2D> texture) override;
+        virtual void setColorAttachment(int index, shared<Texture2D> texture) override;
         virtual Vector2 getColorAttachmentSize() const override;
-        virtual size_t getColorAttachmentCount() const override;
-        virtual shared<Texture2D> getColorAttachment(size_t index) const override;
+        virtual int getColorAttachmentCount() const override;
+        virtual shared<Texture2D> getColorAttachment(int index) const override;
 
     private:
         friend class RenderTarget;
@@ -29,7 +29,7 @@ namespace solo
 
         GLuint handle = 0;
         GLuint depthBufferHandle = 0;
-        std::unordered_map<size_t, shared<Texture2D>> colorAttachments;
+        std::unordered_map<int, shared<Texture2D>> colorAttachments;
         Vector2 colorAttachmentSize;
         bool anyColorAttachments = false;
     };
