+++ src/platform/stub/SoloStubMesh.cpp
@@ -3,15 +3,15 @@
 using namespace solo;
 
 
-unsigned StubMesh::addIndex(MeshIndexFormat indexFormat)
+int StubMesh::addIndex(MeshIndexFormat indexFormat)
 {
-    indexes.push_back(indexes.size());
+    indexes.push_back(static_cast<int>(indexes.size()));
     indexPrimitiveTypes.push_back(MeshPrimitiveType::Triangles);
-    return indexes.size() - 1;
+    return static_cast<int>(indexes.size() - 1);
 }
 
 
-void StubMesh::removeIndex(unsigned index)
+void StubMesh::removeIndex(int index)
 {
     indexes.erase(indexes.begin() + index);
     indexPrimitiveTypes.erase(indexPrimitiveTypes.begin() + index);
