+++ src/platform/stub/SoloStubRenderTarget.h
@@ -10,16 +10,19 @@ namespace solo
         virtual void bind() override {}
         virtual void unbind() override {}
 
-        virtual void setColorAttachment(size_t index, shared<Texture2D> texture) override {}
+        virtual void setColorAttachment(int index, shared<Texture2D> texture) override {}
+
         virtual Vector2 getColorAttachmentSize() const override
         {
             return Vector2();
         }
-        virtual size_t getColorAttachmentCount() const override
+
+        virtual int getColorAttachmentCount() const override
         {
             return 0;
         }
-        virtual shared<Texture2D> getColorAttachment(size_t index) const override
+
+        virtual shared<Texture2D> getColorAttachment(int index) const override
         {
             return nullptr;
         }
